Move common form handler functionalty to abstract classes
