+++ src/ClientBundle/Action/Add.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace CSBill\ClientBundle\Action;
 
-use CSBill\ClientBundle\Form\Handler\ClientFormHandler;
+use CSBill\ClientBundle\Form\Handler\ClientCreateFormHandler;
 use CSBill\ClientBundle\Entity\Client;
 use SolidWorx\FormHandler\FormHandler;
 use SolidWorx\FormHandler\FormRequest;
@@ -41,6 +41,6 @@ final class Add
      */
     public function __invoke(Request $request): FormRequest
     {
-        return $this->handler->handle(ClientFormHandler::class, new Client());
+        return $this->handler->handle(ClientCreateFormHandler::class, new Client());
     }
 }
